feat: added new plays as well as several changes to the plays behavior
Now there are several plays that have a conditional execution based on a
new variable called mastodon_generate_secrets, which controls whether the
plays responsible for generating secrets are executed or not.
This is done to avoid setting incorrect states due to secret management.
By default the variable is set to false, if the user wants to activate
the plays, an extra var needs to be passed when executing the playbook as
--extra-vars '{ mastodon_generate_secrets: True }'.