feat: added initial configuration for the mastodon deployment
With the current state of the scripts, users can now successfully launch an automatic installation.

There's a lot of improvements that can be done still, like trying to use Ansible modules for most of the
functionality. If not achieved this way though, other measures will be implemented to avoid failed repeated
installations.